chore: move types to another folder
