+++ packages/frontend/src/components/common/ResourceList/ResourceTypeUtils.ts
@@ -1,0 +1,4 @@
+import { DashboardBasicDetails, SpaceQuery } from '@lightdash/common';
+
+export type AcceptedResources = SpaceQuery | DashboardBasicDetails;
+export type AcceptedResourceTypes = 'chart' | 'dashboard';
